Made base quiz functionality work Also fixed quiz files

+++ QuizTaker.h
@@ -38,7 +38,7 @@ int QuizTaker::takeQuiz() {
         }
         std::cout << std::endl;
     }
-    int percent = static_cast<int>(score / questions.size() * 100);
+    int percent = static_cast<int>(((double)score / questions.size()) * 100);
     std::cout << "Your final score: " << percent << " percent (" << score << "/" << questions.size() << ")" <<'\n';
 }
